refactor(tests): F033 split EditorLogic suites
